daje

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: lpicciri <lpicciri@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/26 18:06:41 by lpicciri          #+#    #+#             */
-/*   Updated: 2023/01/26 18:06:42 by lpicciri         ###   ########.fr       */
+/*   Updated: 2023/09/21 23:06:13 by lpicciri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
